💄 style: 优化 manifest 预览的尺寸
