+++ src/pages/chat/features/Header/PluginTag/index.tsx
@@ -1,50 +1,13 @@
-import { ActionIcon, Avatar, Icon, Tag } from '@lobehub/ui';
+import { Avatar, Icon, Tag } from '@lobehub/ui';
 import type { MenuProps } from 'antd';
-import { Badge, Dropdown } from 'antd';
+import { Dropdown } from 'antd';
 import isEqual from 'fast-deep-equal';
-import { LucideRotateCw, LucideToyBrick } from 'lucide-react';
-import { memo, useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Flexbox } from 'react-layout-kit';
+import { LucideToyBrick } from 'lucide-react';
+import { memo } from 'react';
 
 import { pluginHelpers, pluginSelectors, usePluginStore } from '@/store/plugin';
 
-const PluginStatus = memo<{ id: string; title?: string }>(({ title, id }) => {
-  const { t } = useTranslation('common');
-  const [status, fetchPluginManifest] = usePluginStore((s) => [
-    pluginSelectors.getPluginManifestLoadingStatus(id)(s),
-    s.fetchPluginManifest,
-  ]);
-
-  const renderStatus = useMemo(() => {
-    switch (status) {
-      case 'loading': {
-        return <Badge color={'blue'} status={'processing'} />;
-      }
-      case 'error': {
-        return (
-          <ActionIcon
-            icon={LucideRotateCw}
-            onClick={() => {
-              fetchPluginManifest(id);
-            }}
-            size={'small'}
-            title={t('retry')}
-          />
-        );
-      }
-      case 'success': {
-        return <Badge status={status} />;
-      }
-    }
-  }, [status]);
-
-  return (
-    <Flexbox gap={12} horizontal justify={'space-between'}>
-      {title} {renderStatus}
-    </Flexbox>
-  );
-});
+import PluginStatus from './PluginStatus';
 
 export interface PluginTagProps {
   plugins: string[];
